split out and move some files
